Previous versions of QtXlsx made use of private, internal classes for zip file handling (QZipReader and QZipWriter). At some point between Qt versions 5.3 and 5.7, the internal API for these classes changed, and it was not possible to build QtXlsx. This version makes use of public classes adapted from internal Qt code (the QtCompress module: see https://github.com/nezticle/qtcompress and https://github.com/aaron-reeves/qtcompress), and is thus compatible once again with all(?) versions of Qt 5.x.
Attempts to use QtCompress as a stand-alone module were unsuccessful (for reasons that I haven't fully explored).  I have now simply included the code directly (see directory '3rdparty'), so Zip file handling is built-in to the QtXlsx module.  While not necessarily an elegant solution, it does work in the limited number of situations that I've attempted.